refactor: mbti 테스트 아이템 정리 및 파일명 변경
